FIx Travis build...

+++ source/OpenBVE/Parsers/Panel/Panel2CfgParser.cs
@@ -129,7 +129,7 @@ namespace OpenBve {
 														}
 														if (Interface.CurrentOptions.EnableBveTsHacks)
 														{
-															switch (PanelCenterY)
+															switch ((int)PanelCenterY)
 															{
 																case 229:
 																	if (PanelBottom == 768 && PanelResolution == 1024)
